Support custom actions in que jobs (#628)
The que integration used a `Transaction#set_action`, which always
overwrites any custom actions set in the job.

If we use `Transaction#set_action_if_nil` it only sets it if
`Appsignal.set_action` is not called in the job itself. This allows for
more flexibility for the user, to customize action names, but also makes
it compatible with our upcoming improve ActiveJob integration, which
runs wrapped by the Que integration. It would have the same problem as
it also calls `Transaction#set_action_if_nil` in the job wrapped by Que.