Rework Resque and Resque ActiveJob integrations (#630)
The Resque integration was not compatible with the new ActiveJob
integration.

## Previous integration problems with ActiveJob integration

When using ActiveJob, but not using the AppSignal Resque plugin, adding
`include ResqueActiveJobPlugin` to the job classes, it would start a
single transaction, but without calling `Appsignal.stop` it wouldn't be
flushed by the extension to the agent and then never arrive at
AppSignal.com. It's needed to call `Appsignal.stop` for the Resque
integration, because from the logs I could tell it was creating a new
process ID for every job, probably creating a new process and quickly
exiting after a job is processed.

If the module was included, it would start two transaction, but then the
ActiveJob wrapper wouldn't be able finish the transaction, because
`Appsignal.stop` had already been called.

## New integration

The new Resque integration extends the `Resque::Job` class to wrap the
`perform` method. This ensures we wrap the entire Resque job and can
call `Appsignal.stop` at the end. If we would continue to wrap it using
the `around_perform_resque_plugin` method we would again call
`Appsignal.stop` before the ActiveJob integration would be closed.

Previous execution order:

- ActiveJob
- around_perform_resque_plugin

New execution order:

- Resque::Job.perform
- ActiveJob

As an added benefit the Resque integration is now active for all jobs by
default, and doesn't need to be manually included in every job or in a
parent job class.

For ActiveJob, queue time support now depends on the main ActiveJob
integration rather than the separate ResqueActiveJobPlugin.

For testing I've added a new job for Resque 1 and 2, so we test both
version and remain compatible.

## Deprecating old integrations

I've deprecated the old integrations and it prints a deprecation_message
on extend for the Resque integration and include for the Resque
ActiveJob integration. It even mentions the exact file and line of the
inclusion.

Added a module level method to the DeprecationMessage util module so
that it can be called without having to be included. I didn't want to
include it in the Resque integration modules that would be included in
app classes.