Add support for Active Job priority
Adapters that use a priority system (Backburner, Delayed::Job and Que,
as Rails shipped adapters), will now report the priority for the job.

I couldn't find any docs for the `queue_with_priority` class method on
Active Job, but I found it defined as part of the
`ActiveJob::QueuePriority` module. The previously listed adapters do
call the `priority` method that this method sets, so it is used.