Track Active Job job metrics (#633)
Like the Sidekiq integration, track metrics automatically for all
ActiveJob integrations.

Track how many jobs are processed and how many fail. We can't
collect much more because fetching those metrics are adapter specific
and should be implemented per adapter.

I wanted to also track how many jobs are "queued" at a certain time with
a "gauge", but we can only track them at the time of queuing, not later
on like we do for Sidekiq with a gauge. So users would see how many jobs
were put on the queue at a certain time, but now how many jobs are
actually in that queue at that time. We'd probably need to implement
that per job system.