Use Sidekiq jid as transaction ID (#640)
For other integrations, such as Rails, we use the request ID or some
other library specific ID to set the transaction ID.

Sidekiq instead generates a new ID with `SecureRandom.uuid`. This makes
it difficult to match the transaction ID from Sidekiq logs or the web UI
to what is reported in the AppSignal UI.

We also filter out the `jid`, as it's part of the `EXCLUDED_JOB_KEYS`
constant in the AppSignal Sidekiq middleware.

I can't find a scenario where the jid isn't reported. It's been part of
Sidekiq since at least version 2. So I haven't implemented a fallback
for when the jid is missing.