Always report Active Job ID (#639)
When a Active Job is wrapped in a adapter integration, such as Sidekiq,
Delayed::Job, Que and Resque, we don't record the Active Job's ID. This
is because a transaction is already started in the adapter. We only
create a new transaction using the Active Job ID for Active Job's
without a adapter integration.

This change adds the `active_job_id` as tag on the sample at all times,
so that it is available with or without an Active Job adapter wrapper.

I've `dup`-ed the `tags` variable, because otherwise the `active_job_id`
tag would also be registered for the metrics being reported, making them
largely unusable as they report metrics per job execution, and not per
queue. Alternatively we can build the "tags" hash twice for transaction
and metric tags, but that seems like it would take more time.

Move the tag building to a separate method in the ActiveJobHelpers so
the `execute` method doesn't become even longer.