Support Delayed::Job jobs without method name
In #637 we got the report that Structs that get enqueued with
`Delayed::Job.enqueue` get an action name of "unknown".

```
Delayed::Job.enqueue job_object
```

I've updated the action name logic to append `#perform` to the job name
as the last possible option for an action name. `#perform` is the method
Delayed::Job will call in such as scenario.

This fixes #637, and eliminates the need to define a `appsignal_name`
method in the job objects that are enqueued with `Delayed::Job.enqueue`.

I haven't found a scenario where the action name as returned by
Delayed::Job in `job.name` is an empty string or `nil`, so I removed
that test case. If it does happen in some scenario I'm not aware of,
those jobs will be reported as `#perform` instead of `unknown`.