Report Active Job metrics per queue and with prio (#643)
Priority is a tag that is not always present on the
`active_job_queue_job_count` metric. This can differ per library and per
job. This makes adding a magic dashboard for it difficult, as metrics
with a variable number of tags is difficult to configure in the graph
definition.

Split out the metric to report jobs with a priority separately, and tag
them separately.

This means that we report a `active_job_queue_job_count` and
`active_job_queue_priority_job_count` for the same job in some
libraries. I haven't found an adapter that only uses priority and not
also the queue. From what I could see Active Job always set a "default"
queue.

I didn't see the need to also report only priorities at the moment, by
reporting a `active_job_priority_job_count` metric. Priorities over
different queues mean different things for their priority.

This split of the metrics makes it easier to supply a graph in the magic
dashboard that always works and doesn't require user config for
priorities to make it work.

For libraries that don't use a priority system, the graph for priorities
will be empty as no metrics are reported for it.