Support old probe constant names (#644)
We moved the Sidekiq and Puma probes to their own files and module in
the Ruby gem in #623. This caused a problem for at least the Sidekiq
probe because in our docs we tell users how to configure it if needed.
https://docs.appsignal.com/ruby/integrations/sidekiq.html#minutely-probe-configuration

To support the old probe location name, use `const_missing` on the hooks
module to return the new constant names for these probes when called by
their old name. When called this way, log and print a deprecation
message telling users to update the way the constant is called. A docs
update is also required for the 2.11.0 release.

When the new constant is returned by `const_missing` another issue
arrived. In #623 the probes were only loaded on installation. Which
break the manual configuration of the probe in environments where
AppSignal is not active and doesn't install the probes automatically.

To fix this, always require the probe files so the constants are
available. This has the same behavior as before, the probes are loaded
(but not installed with `Appsignal::Minutely.probes.register`) on
AppSignal Ruby gem require.

Also update the API status of the probes classes. The class themselves
aren't private, as we tell users how to customize their configuration,
but all other public methods that are used by the Minutely probe system
remain private.