Use extend instead of method aliasing for testing
In the AppSignal testing "module" we use for testing, extend the
`Appsignal::Transaction` class instead of aliasing the methods. This way
we can call `super` to call the original method instead of calling the
aliased `original_*` method. Less moving parts, which is better.

We can't do the same for the `Appsignal::Extension::Transaction` class
because we need to override the instance methods. For that we need to
use `Module.prepend`, which is only available in Ruby 2.0+, and that
would break the Ruby 1.9 build. So to change that we need to wait to
drop Ruby 1.9.