Print error of AppSignal extension not loaded
When the gem's extension fails to install, we don't fail the
installation because we don't want AppSignal to prevent an app from
deploying or running even if it hasn't installed properly.

However, we do not show this to users in any way once the app has
started. The only way for users to find out is an absence of data being
reported, which is not a great indication, or to look for an
"appsignal.log" file which most people don't know exists.

To alert users that something went wrong, not only log the error, but
print it to STDERR, the warning output of the program. This way new
users will see when running "appsignal install" for the first time. And
existing users can see it when the app is booting.

Hopefully this also makes support a bit easier as users notice issues
sooner.

The app is not negatively affected by printing this error, but the issue
will become more visible to users.

I've tried to find a way to `gem install appsignal` to print a warning
or something when it fails to install, but rubygems seems to hide all
the output from STDOUT and STDERR. So we have to wait until the app has
started to print the error.