Fix configuration for working_directory_path (#653)
It doesn't scan the environment for the key so if set via the env var
the value is ignored. So the env var key option was broken, only methods
that worked were the initial config and config file.

Add the key to the string type keys list so it's loaded properly and add
a test for the scenario.