Ensure config env keys are configured (#654)
As noticed in #653, we didn't configure environment variable versions of
all the config keys in the `load_from_environment` method.

Move the key definitions to constants on the class, so they're easier to
define without increasing method length and access in tests.

I've written one key per line so the definitions are easier to read and
edit than the multiple keys per line approach.

I wrote a test that checks all the defined keys and if the list is the
exact same list as all the defined `ENV_*_KEYS` lists. If they do not
match, because a key is missing or is duplicated, it will fail the
test.

While writing this test I noticed no `APPSIGNAL_DNS_SERVERS` key was
configured, so that's missing as well like
`APPSIGNAL_WORKING_DIRECTORY_PATH` in #653. Added it in this PR. We can
backport it to the main branch if someone runs into it as well,
otherwise include the fix in version 2.11.0.