Proof of concept for fetching Puma metrics through the control app
Currently we try to hook into Puma through a plugin, however
since this plugin runs on the "main" thread, when deploying a new
application and reloading Puma, AppSignal in the main thread isn't reloaded.

In theory this can cause a boot fight, where the agent booted from the main thread
differs from the worker threads with a newer AppSignal version,
see: https://github.com/appsignal/appsignal-ruby/issues/575
for more information.

Instead of trying to hook into Puma automagically, this pull changes
the behavior of the minutely probe and needs to be manually configured.

```ruby

Appsignal::Minutely.probes.register(
  :puma,
  Appsignal::Probes::PumaProbe(:path => "unix://./tmp/pids/pumactl.sock", :auth_token => "123")
)
```

It also requries the control app to be enabled in the Puma config.rb:

```ruby

activate_control_app "unix://./tmp/pids/pumactl.sock", { auth_token: "123" }

```

The metrics emitted here are the same as we collected in the main thread,