Conditionally load the correct ActiveSupport instrument override
Depending on the Ruby version `start`/`finish` or even `finish_with_state` is defined
in favor of `instrument`.

This commit updates the `install` method for ActiveSupport notifications
to conditionally load the required overrides.

Note that using start/finish behaves slightly differently from `instrument`
as `instrument` rescues any errors in the given block and ensures
the transaction is closed after the error. With start/finish you are not
able to replicate the same behavior as `finish` might never be called
if an error occurs between the `start` and `finish` calls.