Support batched Shoryuken workers
When a worker would receive a batch of messages it would try to
instrument it as a single message. The `sqs_msg` and the `body`
parameters would be Arrays instead of single objects.
This caused the error reported in #661.

To solve this, parse the collection of messages and message bodies as
Arrays when a batch is detected.

Only store the metadata from the first/oldest message, because we can't
store them all as metadata, they would overwrite each other.

The message bodies are stored on the transaction as a Hash with every
key being the message id of the message part of the batch. This way
people can tell apart which parameters belong to which message.

```
{
  "msg1" => "params of message 1", # String or JSON parsed hash
  "msg2" => "params of message 2"
}
```

For non-batched message the `message_id` is added to the transaction
metadata, but not for batched messages because we can't store multiple
values for the `message_id` metadata.