Forward keyword arguments on Ruby 2.7 and 3.0 (#693)
Cherry-picked #692 on main branch.

Update Object instrumentation for Ruby 2.7 and 3.0 so that it doesn't
print deprecation warnings on Ruby 2.7 and raises an error on Ruby 3.0.

It didn't look broken because we didn't actually test argument
forwarding so the tests didn't break for Ruby 3.0.

I had to implement it twice, once for Ruby 2.7 and higher, and once for
everything older. This is because the two instrumentation methods are
incompatible and this is the easiest way to instrument all versions.
This means two implementation files and two tests because Ruby 1.9 can't
parse the keyword argument syntax introduced in Ruby 2.0. This is an
ugly fix, but it's also present for a very short time, since our develop
branch already drops support for Ruby 1.9 and we can remove this
workaround when this gets merged into the develop branch.
I justed wanted to get this merged into the main branch so we can fully
support Ruby 3.0 in the Ruby gem 2.x series, rather than only fully
support it in the Ruby gem 3.x series.

This change should make our Ruby 3.0 support complete.

Fixes #656