Move Sidekiq middleware to Integrations module
Move the SidekiqPlugin to the Integrations module instead of the Hooks.
This way it follows the pattern of most other integrations.

Rename the SidekiqPlugin to SidekiqMiddleware and set up an
const_missing for it with a deprecation message so the old constant
doesn't break. This is based on the previously implementation that
existed there to deprecate the old SidekiqProbe and PumaProbe constants.