Add Sidekiq error handler
Since Sidekiq 3 there's an error handler hook to listen to errors
reported inside jobs and by Sidekiq internals.

For example, when Sidekiq can't read a malformed JSON when fetching a
job from Redis, it will raise an error before the job and the AppSignal
middleware have started. When this happens, log an error on the
"SidekiqInternal" action and add the internal sidekiq message as the
"sidekiq_error" metadata key.

When Sidekiq encounters an error during a job, it will both raise the
error for the middleware and later in the error handler. To not report
errors double, only set the error in the Sidekiq error handler
middleware. The error is no longer set on the transaction by the
middleware if the middleware encounters it.

When the middleware encounters an error it also doesn't complete the
transaction, but instead lets the error handler close the transaction
later on. If the middleware would already complete it there would be no
current transaction in the error handler anymore.
We only complete the transaction from the middleware when it doesn't
encounter an error and needs to complete it for a performance
measurement.

Because we no longer close the transaction in the middleware on error,
but in the error handler, we only support Sidekiq 3.0 from this change
on. Sidekiq 3.0 was released on March 28, 2014.