Prepend Sidekiq middleware (#698)
Use `prepend` to ensure that the AppSignal middleware is the first
middleware in the chain. This way the AppSignal middleware will also
wrap any other middleware in the chain, and we will catch more errors
this way, but also provide instrumentation that includes the duration
from other middleware for performance samples.

This can still "break", where AppSignal is not the first middleware in
the chain. If another integration or the app itself calls `prepend` as
well, the last caller will be the first in the chain.

I added a small fallback on using `add` when `prepend` is not available
in the Sidekiq middleware chain. It was added in Sidekiq 3.3.0.

Closes #526