Require Ruby 2.0 for gem
Now that we have dropped Ruby 1.9 support, bump the required Ruby
version in the gemspec. We cannot guarantee it will work on Ruby 1.9.

Most integrations using the `Module.prepend` patching style wil not work
as `Module.prepend` was introduced in Ruby 2.0. The object
instrumentation also no longer is compatible with Ruby 1.9 as it has
explicit support of keyword arguments now, also introduced in Ruby 2.0.

Safe to say there is no way the gem will work on Ruby 1.9 any more, so
limit the Ruby version it can be installed on by updating the gemspec.