Support set_error to yield the transaction
This allows users to add more metadata in the `set_error` helper and
makes it behavior similar to `send_error`. This gives users an easy way
to add more metadata to the transaction in the same block.

```
Appsignal.set_error(StandardError.new("my_error")) do |transaction|
  transaction.set_action("my_action")
  transaction.set_namespace("my_namespace")
end
```

Similar change to #481.

With this change we can deprecate the `tags` and `namespace` arguments
as per #678.