Deprecate set_error tags and namespace arguments
Favor the block method of setting the tags and namespace for the
transaction created for the to be set error. The advantage is that this
method is more flexible if more data than tags and namespaces need to be
set, as any method on the transaction object can be called.

This method has our preference moving forward. It works similar in the
Elixir integration, giving along a function (instead of a block) to set
these things on spans.

Part of and closes #678