Update listen_for_error helper send_error usage
Update the `listen_for_error` helper to use the block method to set
metadata (tags and namespace) on the transaction.

In the previous commit calling the `send_error` helper with tags and
namespace as arguments was deprecated and printed and logged a warning,
so update the method call to not print the deprecation warning.