Mark minutely probe thread as fork-safe
The minutely probe thread is designed to live in the main Puma process
and not exist in worker processes but triggers Puma's warning message
about existing threads because of this design.

In rails/rails#40399 a method for marking such threads as fork-safe
for multi-threaded application server (e.g. Puma) was established
by setting a thread local variable called `:fork_safe` to true.