Fix ActionCable integration in test environment
When an Rails 6+ app test suite calls `stub_connection` for ActionCable
tests in an app AppSignal integrates with, it would fail with the
following error as reported in #684:

```
undefined method `env' for #<ActionCable::Channel::ConnectionStub>
```

To fix this, check if the `env` method is available before calling it to
prevent the error from occurring. Ideally we fix this upstream and make
the `ConnectionStub` behave more like an actual connection but with this
fix it at least works from our end now.

This change breaks on Rails 5 as there is no
`action_cable/channel/test_case` file/helper, it appears to use the real
connection class in that version.