Remove middleware insert for JS middleware (#711)
This line was change in PR #686, but hastily done by me. I removed the
second argument from this command, but I should have removed the entire
command as this code now causes an `ArgumentError` error on Rails app
boot.

With config:

```
# config/appsignal.yml
production:
  enable_frontend_error_catching: true
```

Causes this error on `rails server`:

```
ArgumentError: wrong number of arguments (given 1, expected 2+)
```

Remove the entire if-statement because the JavaScript error catcher
middleware and options have been removed in favor of the JavaScript
front-end integration.