Fix failing extension test setup
If the following command is run from the RSpec process the extension is
already loaded and the test does not accurately test what happens when
the extension is not loaded.

```
_TEST_APPSIGNAL_EXTENSION_FAILURE=true rake extension:install
```

Move this command to the Rake task `rake test:failure` to be executed
before RSpec is run and AppSignal is loaded, where it tries to load the
extension.

This will result in more accurate test results.

Print warnings for either scenario (the extension is loaded or is not
loaded) where the opposite state of the extension is detected. This
should help with debugging when switching between two specs where the
different extension installation results are tested.