Lock gem dependencies in build for old Rubies
Similar to #718 and 989ac0d13592f1c0022efa69a503f4e68ffe2469
where this same fix was made for the rexml gem.

I've deleted all the caches because it got random when a job failed or
not. Then wen through all the issues with mismatched gems and gems that
do no longer work on older Ruby versions.

The fix for the `public_suffix` gem, by adding the version lock on the
gem in the gemspec, is not the ideal solution. I would have preferred to
add it the independent gemfiles in the `gemfiles/` directory. However,
that requires a lot of copy paste that becomes very brittle over time,
and would be required for every gemfile and any new gemfiles.

Remove a lot of gems from the build matrix for Ruby 2.0 because all the
exceptions make it impossible to resolve it on this old Ruby version.

## Alternative solution

I wanted to move this all to the `Gemfile` in the root, but this caused
issues with gem load order in the test suite and caused all Rails specs
to fail (Rails was loaded after the AppSignal gem and was not detected).

```
# gemfiles/no_dependencies.gemfile
eval_gemfile("../Gemfile")

gem "rack", "~> 1.6"
```

```
# Gemfile
source "https://rubygems.org"

gemspec
if Gem::Version.new(RUBY_VERSION) < Gem::Version.new("2.3.0")
  # public_suffix 4.0 and newer don't support older Ruby versions
  gem.add_development_dependency "public_suffix", "< 4.0"
end
```

## Registered dependency?

The public_suffix gem doesn't become a registered development dependency (on
RubyGems.org) as long as the gem is published from a Ruby that's newer than
Ruby 2.3, which we usually do. This fix is really only for the Ruby 2.2 and
lower gem builds on the CI. Even if it would be published, it's only a
development dependency and our users would not be affected.