Add mono config
Add a mono config file for use with the [mono
project](https://github.com/appsignal/mono/).

I didn't add mono to the CI test setup as the Ruby gem's build matrix is
very different. We customize the versions of RubyGems and Bundler in the
current build matrix with a custom "bundler_wrapper" script. We can also
see if we can add it later, but the benefits of adding it to the CI are
more present for mono repos.