Refactor build matrix setup
Move the global env vars, prologue and epilogue to the build matrix
itself. We want to add a diagnose job that doesn't need all the env
vars, and prologue and epilogue setup. That actually causes problems
with the expectation that everything has a gemset and Gemfile specified.

This creates a more verbose `.semaphore.yml` file, but it's good enough
for now. It's not the expectation that anyone reads that file, but reads
the `build_matrix.yml` file instead.