Skip empty HTTP proxy during extension download
If the HTTP proxy config was set to an empty String during the AppSignal
extension installation the following error would be logged in the
`install.report` file (available through the diagnose) for the download
step:

```
RuntimeError: Non-HTTP proxy URI:
```
(It is trying to display an empty String in this error message.)

To fix this, when any of the HTTP proxy config returns an empty string,
skip that config value. Only when a config option returns a non-empty
String and non-nil value, use that as the config value. This should fix
download issues where an empty String is used as a HTTP proxy.

This can happen when one of the following config methods are configured
to an empty String, and probably other methods I have not tested as
well.

- `http_proxy=""` env var
- `http_proxy=" "` env var
- `HTTP_PROXY=""` env var
- `HTTP_PROXY=" "` env var
- `~/.gemrc` contains the following config: `http_proxy: ""`
- `~/.gemrc` contains the following config: `http_proxy: " "`