Send Puma metrics to AppSignal agent StatsD server
The Puma plugin causes issues with phased restarts as reported in #575.
The report said that after a phased restart the AppSignal gem would stop
reporting data.

While I couldn't reproduce this particular issue, we did research
potential causes and explore different ways to fetch and send this data,
and found some other issues along the way.

A potential problem is that the AppSignal gem is loaded in the Puma main
process, along with its minutely probes thread, and this can cause
issues with code reloading between (phased) restarts. Would the Ruby gem
be updated the AppSignal code would not be updated because it's loaded
and running in the main process.

## Current situation

Currently the AppSignal Puma plugin starts the minutely probe thread in
the Puma main process and collects the stats there (PR #570). The
original implementation started this minutely probe thread in every
worker (PR #488).

We moved the minutely thread to the main process because in the Puma
clustered mode not all the stats were available and would be reported as
0.

Another issue was that running the probes in every worker would mean the
probe would run as many times as there were workers, sending the data
multiple times. Since all the metrics were gauges this wouldn't be a
problem, but would limit us for future metrics to just this metric.
This problem also exists for all other probes.

## Proposed solution

The proposed solution doesn't rely on the AppSignal gem to send the
metrics data to the extension and finally the agent. Instead it sends
the metrics directly to the AppSignal agent's StatsD server. This first
requires the StatsD server to be enabled, through the private env var in
`config.rb`.

This way we don't need to start the AppSignal gem in the main process.
It is still loaded but not started. We also don't boot another thread
for minutely probes. It's a more lightweight solution.

Log all errors as normal logs, because error logs stop the Puma process.

### About prune_bundler

Since the gem is still loaded this can cause issues with the
`prune_bundler` option, the Puma option to unload all gems from the main
process. This makes the AppSignal plugin unavailable, and requires the
following option to be set to load it again so the plugin can be found.

```
# config/puma.rb
prune_bundler
extra_runtime_dependencies ["appsignal"]
```

Because this loads the AppSignal gem in memory again it may cause some
the issues mentioned earlier again. If this becomes a real problem we
can also decide to extract this plugin into its own separate gem to
be able to load it separately.

## Communication with Puma control server

In PR #655 Robert explored communicating with the Puma control server
over a socket to fetch the data from the server's `/stats` endpoint.
This was needed because it moved the Puma probe back into the workers,
where `Puma.stats` does not have the metrics available.

This solution also required an additional gem to send HTTP requests over
the socket. We're trying to limit the number of dependencies the gem
relies on, currently we only have a dependency on `rack`, and we don't
want to add this dependency only for the Puma integration.

If the solution in this commit does not work well enough in the end, and
we don't want to create a separate gem for the Puma plugin, and we find
another implementation to communicate with the Puma control server via
the socket I suggest we try this solution next.