Stabilize puma plugin specs
The timings were not perfect for these specs and they failed randomly on
the CI. Locally it took me 236 automated retries for the specs to fail finally.

To make it more stable, use the `wait_for` helper to wait until the data
has come in. It waits 5 seconds maximum so if it fails it wouldn't take
forever to fail.