Add Linux ARM flag to enable experimental build (#731)
Users can enable the Linux ARM build with the
`APPSIGNAL_BUILD_FOR_LINUX_ARM` environment variable during
installation. This build is by default behind a flag because it requires
more testing. People may expect it just works because it gets chosen
automatically if we don't put it behind a flag. Now we have an
opportunity to warn them beforehand.

I moved the `agent_architecture` logic to its own method on the System
module so we can reuse it more easily.