Check Rails.backtrace_cleaner method exists
Check if the Rails.backtrace_cleaner method exists before calling it.
We've got a report from a user that the method call fails with a
NoMethodError in a Sinatra app using the activesupport and actionmailer
gems version 5.2.0.

```
NoMethodError - undefined method `backtrace_cleaner' for Rails:Module:
/gems/appsignal-3.0.6/lib/appsignal/transaction.rb:530:in `cleaned_backtrace'
```

I was unable to reproduce in the test setups, but I can see a scenario
where this could fail. To make sure we don't cause this kind of error in
such a scenario, check if the Rails module listens to the
`backtrace_cleaner` method before calling it.

https://github.com/appsignal/test-setups/commit/d484b3676c7f5c0ee9bc54e99c662e783028cc12