Fix extension installation failure NoMethodError
Backport of #720

## Scenario

When the extension installation fails, the AppSignal Ruby gem should
load but not cause any errors that cause the app to stop whenever an
extension method is called.

The extension always fails on Microsoft Windows and Apple M1 machines
currently, but it can also fail for various other reasons, on systems we
expect it to work.

In issue #707, later documented in #719 in more detail, a scenario was
described where a user would call the `Appsignal.increment_counter`
method and it caused the following error.
#707 (comment)

```ruby
# In Ruby it trips on adding tags to a data object
Appsignal.increment_counter "foo", 1, { "foo" => "bar" }
# In JRuby this is enough, just creating the data object for an empty hash
Appsignal.increment_counter "foo", 1
```

Results in:

```
/integration/lib/appsignal/utils/data.rb:24:in `block in map_hash': undefined method `set_string' for nil:NilClass (NoMethodError)
  from /integration/lib/appsignal/utils/data.rb:20:in `each'
  from /integration/lib/appsignal/utils/data.rb:20:in `map_hash'
  from /integration/lib/appsignal/utils/data.rb:10:in `generate'
  from /integration/lib/appsignal/helpers/metrics.rb:35:in `increment_counter'
  from app.rb:29:in `perform'
  from /integration/lib/appsignal/integrations/object.rb:32:in `block (2 levels) in appsignal_instrument_method'
  from /integration/lib/appsignal/helpers/instrumentation.rb:525:in `instrument'
  from /integration/lib/appsignal/integrations/object.rb:31:in `block in appsignal_instrument_method'
  from app.rb:36:in `block (2 levels) in <main>'
  from /integration/lib/appsignal/helpers/instrumentation.rb:77:in `monitor_transaction'
  from app.rb:35:in `block in <main>'
  from app.rb:34:in `times'
  from app.rb:34:in `<main>'
```

This was caused by the `Appsignal::Extension.data_map_new` and
`Appsignal::Extension.data_array_new` methods returning nil, and then
the `Appsignal::Data`'s `map_hash` and `map_array` methods calling
various `set_*` methods on that `nil` object.

```
map = Appsignal::Extension.data_map_new
# map = nil
map.set_string(key, value)
# NoMethodError - undefined method `set_string' for nil:NilClass
```

The same problem occurs for Transactions being created in the following
way, because the `@ext`, transaction extension object, is `nil` in the
`Appsignal::Transaction` object and any method calls on `nil` will raise
a `NoMethodError`.

```
transaction = Appsignal::Transaction.create(
  "demo-id",
  Appsignal::Transaction::HTTP_REQUEST,
  Appsignal::Transaction::GenericRequest.new({})
)
Monitor.new.perform("foo", { :foo => i }, foo: i, bar: i.even?)
Appsignal::Transaction.complete_current!
```

Results in:

```
/integration/lib/appsignal/transaction.rb:349:in `finish_event': undefined method `finish_event' for nil:NilClass (NoMethodError)
  from /integration/lib/appsignal/helpers/instrumentation.rb:529:in `ensure in instrument'
  from /integration/lib/appsignal/helpers/instrumentation.rb:529:in `instrument'
  from /integration/lib/appsignal/integrations/object.rb:31:in `block in appsignal_instrument_method'
  from app.rb:44:in `block in <main>'
  from app.rb:34:in `times'
  from app.rb:34:in `<main>'
/integration/lib/appsignal/transaction.rb:344:in `start_event': undefined method `start_event' for nil:NilClass (NoMethodError)
  from /integration/lib/appsignal/helpers/instrumentation.rb:524:in `instrument'
  from /integration/lib/appsignal/integrations/object.rb:31:in `block in appsignal_instrument_method'
  from app.rb:44:in `block in <main>'
  from app.rb:34:in `times'
  from app.rb:34:in `<main>'
```

## Solution

I added `Appsignal::Extension::MockData` and
`Appsignal::Extension::MockTransaction` classes to act as a fallback for
when the extension is not loaded. Like the `Appsignal::Extension` class
this implements a `method_missing` method that captures any method calls
and will then do nothing. It will prevent any `NoMethodError` from being
raised when any methods are called on the data object

This `MockData` inherits from `Data` so it will pass any type checks in
other methods, like in the JRuby extension.

## Alternative solution

I also considered checking if the extension was loaded in the `map_hash`
and `map_array` methods , and calling `return` early to skip all method
calls to the (not loaded) extension.

This would require any part of the code that interacts with the
extension to have such a guard for that condition. Which adds a lot of
extra checks, which are easy to forget, causing similar issues to be
easily introduced.

It would also cause issues where the return value was interacted with,
such as in the JRuby extension wrapper in `finish_event`. Here it does a
type check and fails when it does not pass it. A `nil` value would fail
that type check and an `ArgumentError` would be raised.