Fix JRuby build by specifying GCC version
Backport of https://github.com/appsignal/appsignal-ruby/pull/737

The JRuby builds were failing because the JRuby installation failed when
switching using `sem-version ruby jruby-<version>`. The installation
failed because we're using a different version than the one installed by
default on the Semaphore VM. The default version was probably updated
and then this broke the build for our setup.

The JRuby builds we use need a newer GCC version than the one selected
by default (GCC 6).  Switch the C language to version 8 to also switch
to GCC 8 and allow JRuby to be installed properly.

Bump the JRuby version as well, we don't really need to test against the
older one anymore, it's EOL.