Print diagnose String values surrounded by quotes
This makes the Ruby integration print the diagnose output consistently
with the Node.js integration: String values are surrounded by quotes
when printed.

I added a new `puts_format` helper method (please rename if you think of
something better), to print the values like `puts_value` but print
Strings with quotes around them. As a shortcut I call `#inspect` on the
object to accomplish this. Other types of values should be printed like
we expect as well.

I split up some calls to `puts_and_save`. It's difficult to save it as
the plain value, but print it as the "inspected" value. Now that I look
at it several years later `puts_and_save` is very easy, but it does a
little bit too much.