Split up filter_data_keys option from Span API
The filter_data_keys config option was previously used in the Span API
to filter out any kind of data object (parameters and session data), but
the merged config values could cause some unexpected behavior, filtering
out more than the user intended.

The config options have been split up in the `filter_parameters` and
`filter_session_data` options. The filtering is still done in the
extension for the Span API, so those two config options also need to be
set in the environment for the extension.

Since the filtering is now done with a data object is given to the
`appsignal_set_span_sample_data` function, the
`appsignal_data_filtered_map_new` function implementation can be
removed.

This change does not affect the default Transaction API.