Use log_level option in Ruby logger
We've added the log_level option in PR #773. This primarily configured
it for the extension and agent, but not the integration's logger itself.

Add a `Appsignal::Config#log_level` helper method to figure out what log
level to set based on deprecated options and the new log_level option.
This mimics the extension and agent's behaviors. It relies on the
extension's output being printed to STDERR for (deprecation) warnings
about the configuration, so no warnings are printed by the Ruby gem
itself.

I've removed the default config of `log_level: info` so that debug and
transaction_debug_mode can still overwrite that level in the Ruby gem.
This is the same behavior as the extension and agent.