Fix test failures for Rails 7
This commit fixes the test failures in Rails 7 that were present
due to changes in the ActiveJob internals.

When a job is performed, the `perform_start.active_job` event now
seems to always fire _after_ the `perform.active_job` event.

When a message is serialised by ActiveJob, the arguments and
params passed to the method are now always wrapped in a separate
hash within the `params` key, itself containing `args` and `params`
keys, and subject to the quirks of ActiveJob hash serialisation.

This commit modifies the tests to expect these changes.