Fix ActiveJob tests for Rails 6.1 and 7.0
Apply the same ActiveJob fixes to the Sidekiq specs. Move it to its own
helper module to reduce duplicate code.

There are some changes between Rails 6.1 and 7.0 in the format of the
arguments. Most of the fixes for Rails 7 can also be applied to Rails
6.1, but only if it's running on Ruby 3.1.0 or newer. It's weird
behavior, hence the weird if statement, but Rails must be doing some
kind of Ruby version check for Ruby 3.1, but only in Rails 6.1.

[skip changeset] as it's only test related changes.